fix: reorganize user-related schemas and update imports for making e-mail adress not updateable.
fix: invalidation now works for user/trainee list/ instructor list
