+++ frontend/src/lib/api/collections/learningKit.ts
@@ -3,7 +3,7 @@ import { UpdateLearningKitSchema } from '$lib/schemas/request/UpdateLearningKit'
 import { LearningKitPageSchema, LearningKitResSchema } from '$lib/schemas/response/LearningKitRes';
 import { createEndpoint } from '../createEndpoint';
 import { z } from 'zod';
-import { CreateTraineeUserSchema } from '$lib/schemas/request/CreateTraineeUser';
+import { CreateTraineeUserSchema } from '$lib/schemas/request/user/CreateTraineeUser';
 import { UpdateLearningUnitOrderSchema } from '$lib/schemas/request/UpdateLearningUnitOrder';
 import { GenericSuccessSchema } from '$lib/schemas/response/GenericSuccess';
