+++ frontend/src/lib/i18n/locales/en.json
@@ -320,5 +320,6 @@
 	"user.form.edit.title": "Edit User",
 	"users.overview.reset.error": "Failed to reset password",
 	"users.overview.reset.loading": "Resetting password...",
-	"users.overview.reset.success": "User password reset successfully"
+	"users.overview.reset.success": "User password reset successfully",
+	"user.form.edit.emailLocked": "Email is locked and cannot be changed"
 }
