+++ frontend/src/lib/schemas/request/user/CreateTraineeUser.ts
