+++ frontend/src/routes/(app)/users/[userId]/edit-form/+page.server.ts
@@ -4,35 +4,38 @@ import { handleApiError } from '$lib/api/apiError';
 import { type Actions, error, fail, redirect } from '@sveltejs/kit';
 import { api } from '$lib/api/apiClient';
 import { getUser, updateUser } from '$lib/api/collections/user';
-import { UserResSchema } from '$lib/schemas/response/UserResSchema';
-import { UpdateUserSchema } from '$lib/schemas/request/UpdateUser';
+import { UpdateUserSchema } from '$lib/schemas/request/user/UpdateUser';
 
 export const load = async ({ fetch, params }) => {
 	if (!params.userId) {
-		return error(400, 'LearningKit ID is required');
+		throw error(400, 'User ID is required');
 	}
 
 	const user = await api(fetch).req(getUser, null, params.userId).parse();
 	const form = await superValidate(
 		{
-			username: user.username,
 			name: user.name,
 			surname: user.surname,
 			role: user.role
 		},
-		zod(UserResSchema)
+		zod(UpdateUserSchema)
 	);
 
-	return { form };
+	return {
+		form,
+		username: user.username
+	};
 };
 
 export const actions = {
 	update: handleApiError(async ({ request, fetch, params }) => {
 		const form = await superValidate(request, zod(UpdateUserSchema));
+
 		if (!form.valid || !params.userId) {
 			return fail(400, { form });
 		}
+
 		await api(fetch).req(updateUser, form.data, params.userId).parse();
-		redirect(303, `/users`);
+		return redirect(303, '/users');
 	})
 } satisfies Actions;
